# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lumino 记账本服务",
        "title": "Lumino",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/account": {
            "get": {
                "tags": [
                    "账户"
                ],
                "summary": "获取账户",
                "parameters": [
                    {
                        "type": "number",
                        "description": "账户余额",
                        "name": "balance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-03-26T00:00:00Z",
                        "description": "DeletedAt 删除时间",
                        "name": "deleted_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户描述",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户图标",
                        "name": "icon",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "账户类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updated_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "创建人",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "账户结果",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Account"
                            }
                        }
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "账户"
                ],
                "summary": "修改账户",
                "parameters": [
                    {
                        "type": "number",
                        "description": "账户余额",
                        "name": "balance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-03-26T00:00:00Z",
                        "description": "DeletedAt 删除时间",
                        "name": "deleted_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户描述",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户图标",
                        "name": "icon",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "账户类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updated_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "创建人",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "账户"
                ],
                "summary": "注册账户",
                "parameters": [
                    {
                        "type": "number",
                        "description": "账户余额",
                        "name": "balance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-03-26T00:00:00Z",
                        "description": "DeletedAt 删除时间",
                        "name": "deleted_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户描述",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户图标",
                        "name": "icon",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "账户类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updated_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "创建人",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "账户"
                ],
                "summary": "删除账户",
                "parameters": [
                    {
                        "type": "number",
                        "description": "账户余额",
                        "name": "balance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建时间",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-03-26T00:00:00Z",
                        "description": "DeletedAt 删除时间",
                        "name": "deleted_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户描述",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户图标",
                        "name": "icon",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "主键id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账户名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "账户类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "更新时间",
                        "name": "updated_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "创建人",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Account": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "账户余额",
                    "type": "number"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "DeletedAt 删除时间",
                    "type": "string",
                    "example": "2025-03-26T00:00:00Z"
                },
                "description": {
                    "description": "账户描述",
                    "type": "string"
                },
                "icon": {
                    "description": "账户图标",
                    "type": "string"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "name": {
                    "description": "账户名称",
                    "type": "string"
                },
                "type": {
                    "description": "账户类型",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "创建人",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}